pyproject.toml
[project]
name = "sentiment-backend"
version = "0.1.0"
description = ""
authors = [
    { name = "goddessana", email = "twicegoddessana1229@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "black (>=25.1.0,<26.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.8.0,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
